+++ test/unit/fuzz-tools.test.js
@@ -142,6 +142,7 @@ import * as pedsGrowthV141 from '../../lib/peds-growth-v141.js';
 import * as surgV142 from '../../lib/surg-v142.js';
 import * as frailtyV143 from '../../lib/frailty-v143.js';
 import * as orthoV144 from '../../lib/ortho-v144.js';
+import * as orthoV145 from '../../lib/ortho-v145.js';
 import * as growthLmsData from '../../lib/growth-lms-data.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
@@ -230,6 +231,7 @@ const MODULES = {
   'surg-v142.js': surgV142,
   'frailty-v143.js': frailtyV143,
   'ortho-v144.js': orthoV144,
+  'ortho-v145.js': orthoV145,
   'growth-lms-data.js': growthLmsData,
   'ems-v149.js': emsV149,
 };